+++ tests/Controllers/PlainWorkerController.php
@@ -1,0 +1,8 @@
+<?php
+
+namespace HassanDomeDenea\HddLaravelHelpers\Tests\Controllers;
+
+class PlainWorkerController extends WorkerController
+{
+    protected array $with = [];
+}