+++ tests/Controllers/WorkerWithPreloadedCategoryController.php
@@ -1,0 +1,10 @@
+<?php
+
+namespace HassanDomeDenea\HddLaravelHelpers\Tests\Controllers;
+
+use HassanDomeDenea\HddLaravelHelpers\Tests\Actions\CreateWorkerWithCategoryLoadedAction;
+
+class WorkerWithPreloadedCategoryController extends WorkerController
+{
+    public ?string $createActionClass = CreateWorkerWithCategoryLoadedAction::class;
+}