+++ tests/Data/Worker/StoreWorkerData.php
@@ -1,0 +1,13 @@
+<?php
+
+namespace HassanDomeDenea\HddLaravelHelpers\Tests\Data\Worker;
+
+use Spatie\LaravelData\Data;
+
+class StoreWorkerData extends Data
+{
+    public function __construct(
+        public string $name,
+        public int $category_id,
+    ) {}
+}