+++ tests/Models/Category.php
@@ -1,0 +1,18 @@
+<?php
+
+namespace HassanDomeDenea\HddLaravelHelpers\Tests\Models;
+
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\HasMany;
+
+class Category extends Model
+{
+    protected $fillable = [
+        'name',
+    ];
+
+    public function workers(): HasMany
+    {
+        return $this->hasMany(Worker::class);
+    }
+}