+++ tests/Models/Tag.php
@@ -1,0 +1,19 @@
+<?php
+
+namespace HassanDomeDenea\HddLaravelHelpers\Tests\Models;
+
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
+
+class Tag extends Model
+{
+    protected $fillable = [
+        'name',
+        'worker_id',
+    ];
+
+    public function worker(): BelongsTo
+    {
+        return $this->belongsTo(Worker::class);
+    }
+}