+++ tests/Models/Worker.php
@@ -1,0 +1,25 @@
+<?php
+
+namespace HassanDomeDenea\HddLaravelHelpers\Tests\Models;
+
+use Illuminate\Database\Eloquent\Model;
+use Illuminate\Database\Eloquent\Relations\BelongsTo;
+use Illuminate\Database\Eloquent\Relations\HasMany;
+
+class Worker extends Model
+{
+    protected $fillable = [
+        'name',
+        'category_id',
+    ];
+
+    public function category(): BelongsTo
+    {
+        return $this->belongsTo(Category::class);
+    }
+
+    public function tags(): HasMany
+    {
+        return $this->hasMany(Tag::class);
+    }
+}